+++ spec/hamli/parser_spec.rb
@@ -354,5 +354,35 @@ RSpec.describe Hamli::Parser do
         )
       end
     end
+
+    context 'with Haml comment' do
+      let(:source) do
+        <<~HAML
+          -# a
+        HAML
+      end
+
+      it 'returns expected S-expression' do
+        is_expected.to eq(
+          [:multi, [:newline]]
+        )
+      end
+    end
+
+    context 'with Haml comment' do
+      let(:source) do
+        <<~HAML
+          -#
+            a
+              b
+        HAML
+      end
+
+      it 'returns expected S-expression' do
+        is_expected.to eq(
+          [:multi, [:newline], [:newline], [:newline]]
+        )
+      end
+    end
   end
 end