+++ lib/hamli/parser.rb
@@ -31,6 +31,8 @@ module Hamli
         parse_html_comment_line
       elsif @scanner.match?(/-#/)
         parse_haml_comment_line
+      elsif @scanner.match?(/=/)
+        parser_output_line
       else
         parse_text_line
       end
@@ -390,6 +392,20 @@ module Hamli
       @scanner.pos += @scanner.matched_size
       end_ = @scanner.charpos
       [:hamli, :interpolate, begin_, end_, value]
+    end
+
+    # Parse output line part.
+    #   e.g. = abc
+    #        ^^^^^
+    def parser_output_line
+      @scanner.pos += 1
+      @scanner.scan(/[ \t]*/)
+      block = [:multi]
+      begin_ = @scanner.charpos
+      content = parse_broken_lines
+      end_ = @scanner.charpos
+      @stacks.last << [:hamli, :position, begin_, end_, [:hamli, :output, content, block]]
+      @stacks << block
     end
 
     # @note Broken line means line-breaked lines, separated by trailing "," or "|".