+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Support output block in tag line.
+- Support HTML comment line.
 
 ## 0.2.0 - 2022-01-25