+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Support HTML comment line.
 - Support HTML conditional comment line.
 - Support Haml comment line.
+- Support output line.
 
 ## 0.2.0 - 2022-01-25